Bump to IceCC 1.2
Update IceCC/Icecream to version 1.2, which is the same version
currently shipped with Debian stable (Buster). This bump is needed
because IceCC 1.1 does not seem to interoperate with the scheduler from
newer versions and it is expected that people who run IceCC will install
IceCC in their build farm from distribution packages.

While at it, specify the Git commit identifier for the 1.2 release
to ensure the reproducibility of the builds.